+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## V1.91
+### script
+* support Home Assistant over HTTPS (https://github.com/reserve85/HoymilesZeroExport/issues/178)
+* Support login credentials for Tasmota (https://github.com/reserve85/HoymilesZeroExport/issues/159)
+### config
+* add `[HOMEASSISTANT]`: `HA_HTTPS` 
+* add `[INTERMEDIATE_HOMEASSISTANT]`: `HA_HTTPS_INTERMEDIATE`
+* add `[TASMOTA]`: `USER` and `PASS` 
+* add `[TASMOTA_INTERMEDIATE]`: `USER_INTERMEDIATE` and `PASS_INTERMEDIATE` 
+
 ## V1.90
 ### script
 * fix HOY_BATTERY_THRESHOLD_NORMAL_LIMIT_IN_V, see https://github.com/reserve85/HoymilesZeroExport/issues/174